+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pipelab/cli
 
+## 2.0.0-beta.28
+
+### Patch Changes
+
+- sd
+- Updated dependencies
+  - @pipelab/constants@1.0.0-beta.24
+  - @pipelab/core-node@1.0.0-beta.28
+  - @pipelab/shared@1.0.0-beta.23
+
 ## 2.0.0-beta.27
 
 ### Patch Changes