+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/test-utils
 
+## 1.0.0-beta.29
+
+### Patch Changes
+
+- sd
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.0-beta.29
+
 ## 1.0.0-beta.28
 
 ### Patch Changes