+++ plugins/plugin-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pipelab/plugin-core
 
+## 1.0.0-beta.29
+
+### Patch Changes
+
+- sd
+- Updated dependencies
+  - @pipelab/core-node@1.0.0-beta.28
+
 ## 1.0.0-beta.28
 
 ### Patch Changes