+++ plugins/plugin-discord/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pipelab/plugin-discord
 
+## 1.0.0-beta.29
+
+### Patch Changes
+
+- sd
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.0-beta.29
+  - @pipelab/shared@1.0.0-beta.23
+
 ## 1.0.0-beta.28
 
 ### Patch Changes